quickstart: split system installation and process invocation into own package
